三国咸话/答题/requirements.txt
requests
pillow
uvicorn
fastapi